Move current_ip into Utils
